Group `Text` align examples
