--- polaris.shopify.com/src/pages/examples/text-with-align-center.tsx
@@ -1,14 +1,0 @@
-import { Text } from "@shopify/polaris";
-import React from "react";
-import { withPolarisExample } from "../../components/PolarisExamplePage";
-
-function TextExample() {
-  return (
-    <Text variant="bodyLg" as="p" alignment="center">
-      Manage your Shopify store on-the-go with real-time notifications, access
-      to your dashboard, and order management, all from your smartphone.
-    </Text>
-  );
-}
-
-export default withPolarisExample(TextExample);
